requirements.txt
tqdm==4.67.1
polars-u64-idx==1.29.0
requests==2.32.3
shapely==2.1.0
pyproj==3.7.1
matplotlib==3.10.1
numpy==2.2.5
more-itertools==10.7.0
scikit-learn==1.6.1
pytest==8.3.5
chardet==5.2.0
rich==14.0.0
logger==1.4
astropy==7.0.1